# repo/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospitality Operations Management System API
  description: Backend API for managing events, resources, reservations, recipes, inventory, entitlements, and operational workflows.
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login or POST /auth/refresh.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            cursor:
              type: string
              nullable: true
            hasMore:
              type: boolean
            total:
              type: integer
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved, in_service, closed]
        budget:
          type: number
        created_by:
          type: string
          format: uuid
        service_windows:
          type: array
          items:
            $ref: '#/components/schemas/ServiceWindow'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceWindow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        label:
          type: string
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events (filtered by role/ownership)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, approved, in_service, closed]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
    post:
      tags: [Events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                budget:
                  type: number
                service_windows:
                  type: array
                  items:
                    type: object
                    properties:
                      start_time:
                        type: string
                        format: date-time
                      end_time:
                        type: string
                        format: date-time
                      label:
                        type: string
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/{id}:
    get:
      tags: [Events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Events]
      summary: Update event (Draft state only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                budget:
                  type: number
                service_windows:
                  type: array
                  items:
                    type: object
                    properties:
                      start_time:
                        type: string
                        format: date-time
                      end_time:
                        type: string
                        format: date-time
                      label:
                        type: string
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error or invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /events/{id}/materials:
    get:
      tags: [Events]
      summary: Get materials list for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Materials list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        quantity:
                          type: number
                        unit:
                          type: string
    put:
      tags: [Events]
      summary: Update materials list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                materials:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      quantity:
                        type: number
                      unit:
                        type: string
      responses:
        '200':
          description: Materials updated
  /events/{id}/resource-requests:
    get:
      tags: [Events]
      summary: List resource requests for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource requests
    post:
      tags: [Events]
      summary: Create a resource request for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_id, quantity]
              properties:
                resource_id:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                notes:
                  type: string
      responses:
        '201':
          description: Resource request created
        '400':
          description: Validation error or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Approvals ─────────────────────────────────────────
...